Change tests of model becuase of changes in the model classes
